Enable BL2 timestamps
Currently, BL2 doesn't output any timestamp in its prints:

INFO:    BL2: DMA    len=1048576(0x100000)
[0.000000]NOTICE:  BL3-1: v1.1(debug):88ccfd9
[0.000000]NOTICE:  BL3-1: Built : 13:16:21, Nov 22 2016
[0.000025]INFO:    BL3-1: Initializing runtime services

The reason is that the Generic Timer counter is started later in
BL31, so that, even if enabled, BL2 timestamps would output 0:

[0.000000]INFO:    BL2: Loading BL3-3
[0.000000]INFO:    BL2: DMA    len=1048576(0x100000)
[0.000000]NOTICE:  BL3-1: v1.1(debug):dcc8dda
[0.000000]NOTICE:  BL3-1: Built : 12:57:04, Dec  5 2016
[0.000025]INFO:    BL3-1: Initializing runtime services

In other words, in the reference configuration, without moving
the counter initialization from BL31 to BL2, the BL2 timestamps
are quite useless.

When counting is started in BL2, the output becomes:

[0.080872]NOTICE:  BL2: Normal boot
[0.095072]NOTICE:  BL2: dst=0x44100000 src=0x8200000 len=524288(0x80000)
[0.112639]NOTICE:  BL3-1: v1.1(debug):707c954
[0.114309]NOTICE:  BL3-1: Built : 10:25:58, Dec  2 2016
[0.183289]INFO:    BL3-1: Initializing runtime services

This patch only enables the timestamping in BL2, independently of
where the Generic Timer counting is started (BL2 or BL31).

This patch should be applied _after_ making sure the Generic Timer
is started in BL2 instead of BL31, which is being solved separately.

Signed-off-by: Eugeniu Rosca <erosca@de.adit-jv.com>
[dirk.behme: Drop usage of rcar_stack_generic_timer in BL2 case
             introduced with Yocto v2.16]
Signed-off-by: Dirk Behme <dirk.behme@de.bosch.com>